assignments-backend/requirements.txt
Django>=5.1,<6.0
djangorestframework
psycopg2-binary
django-cors-headers